pyproject.toml
[project]
name = "lemmatizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "blis==1.3.0",
    "catalogue==2.0.10",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudpathlib==0.21.1",
    "confection==0.1.5",
    "cymem==2.0.11",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "it-core-news-sm",
    "jinja2==3.1.6",
    "langcodes==3.5.0",
    "langdetect>=1.0.9",
    "language-data==1.3.0",
    "marisa-trie==1.2.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "murmurhash==1.0.13",
    "mysql-connector-python>=9.3.0",
    "numpy==2.2.6",
    "packaging==25.0",
    "passlib[bcrypt]>=1.7.4",
    "pluggy==1.6.0",
    "preshed==3.0.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyjwt>=2.10.1",
    "pytest==8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests==2.32.4",
    "rich==14.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.3.0.post1",
    "sniffio==1.3.1",
    "spacy==3.8.7",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "sqlmodel>=0.0.24",
    "srsly==2.5.1",
    "starlette==0.46.2",
    "thinc==8.3.6",
    "tomli==2.2.1",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "wrapt==1.17.2",
]

[tool.uv.sources]
it-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.8.0/it_core_news_sm-3.8.0-py3-none-any.whl" }
